Interface/AddOns/BetterWardrobe/Libs/LibArmorToken-1.0/tools/requirements.txt
PyYAML==3.11
requests==2.11.1
requests-cache==0.4.12
